add shared `fix' point function for `Lazy'...
This function is a version that uses `Lazy' in its implementation so that it enjoys the recursive evaluation checks introduced with the previous changes.  It forms a recursion on a binding, which can also otherwise be done with a simple anonymous function but without the checks.

This version does not require a recursion limit condition internal to the function argument as the deferred execution ensures that it can not run away indefinitely.